sms: remove superfluous types
